+++ internal/command/jsonplan/values.go
@@ -87,7 +87,7 @@ func marshalPlannedOutputs(changes *plans.Changes) (map[string]Output, error) {
 		ret[oc.Addr.OutputValue.Name] = Output{
 			Value:     json.RawMessage(after),
 			Type:      json.RawMessage(afterType),
-			Sensitive: oc.Sensitive,
+			Sensitive: oc.AfterSensitive,
 		}
 	}